be more careful when `opam switch remove` fails; fixes #8
Users could get in an inconsistent state after running
  opam compiler-conf uninstall
if they answer "no" to one of the interactive yes/no prompts of opam:
the present script would remove its compiler description and consider
the switch removed, which made further compiler-conf install or
uninstall command fail.

It would be nice if $(opam switch remove) returned a different return
code depending on whether the switch was really uninstalled. It
doesn't, so as a workaround we test for the presence of the
switch-specific $PREFIX directory to know whether the uninstall
worked.