stop relying on escaped characters interpretation for portability
OSX does not know of (echo -e) (see #13); I tried using (printf "%b")
instead but it eats the space on my machine. Worked around the whole
issue by turning the OPAM_COMP_DATA variable into a function that
directly prints to stdout (instead of building a string in such a way
that printing it works).